Break up cache serializers
