+++ lib/cache-serializer-factory.js
@@ -74,9 +74,7 @@
  * @returns {Promise} promise that resolves when writing completes
  */
 
-var join = require('path').join;
-
-var FileSerializer = require('./cache-serializers').FileSerializer;
+var FileSerializerPlugin = require('./hard-source-file-serializer-plugin');
 var AppendSerializerPlugin = require('./hard-source-append-serializer-plugin');
 
 /**
@@ -101,15 +99,10 @@ function CacheSerializerFactory(compiler) {
       // Otherwise lets return the default serializers.
       switch (info.type) {
       case 'data':
-        return AppendSerializerPlugin.createSerializer(info);
-        return new LevelDbSerializer({
-          cacheDirPath: join(info.cacheDirPath, info.name)
-        });
+        return CacheSerializerFactory.dataSerializer.createSerializer(info);
         break;
       case 'file':
-        return new FileSerializer({
-          cacheDirPath: join(info.cacheDirPath, info.name)
-        });
+        return CacheSerializerFactory.fileSerializer.createSerializer(info);
         break;
       default:
         throw new Error(
@@ -140,5 +133,15 @@ CacheSerializerFactory.prototype.create = function(info) {
 
   return serializer;
 };
+
+/**
+ * The default data serializer factory.
+ */
+CacheSerializerFactory.dataSerializer = AppendSerializerPlugin;
+
+/**
+ * The default file serializer factory.
+ */
+CacheSerializerFactory.fileSerializer = FileSerializerPlugin;
 
 module.exports = CacheSerializerFactory;
