+++ lib/hard-source-file-serializer.js
@@ -1,0 +1,50 @@
+var fs = require('fs');
+var join = require('path').join;
+
+var _mkdirp = require('mkdirp');
+
+var Promise = require('bluebird');
+
+var mkdirp = Promise.promisify(_mkdirp);
+var fsReadFile = Promise.promisify(fs.readFile, {context: fs});
+var fsReaddir = Promise.promisify(fs.readdir, {context: fs});
+var fsWriteFile = Promise.promisify(fs.writeFile, {context: fs});
+
+module.exports = FileSerializer;
+
+function FileSerializer(options) {
+  this.path = options.cacheDirPath;
+}
+
+FileSerializer.prototype.read = function() {
+  var assets = {};
+  var cacheAssetDirPath = this.path;
+  return mkdirp(cacheAssetDirPath)
+  .then(function() {
+    return fsReaddir(cacheAssetDirPath);
+  })
+  .then(function(dir) {
+    return dir.map(function(name) {
+      return Promise.all([name, fsReadFile(join(cacheAssetDirPath, name))]);
+    });
+  })
+  .then(Promise.all)
+  .then(function(_assets) {
+    for (var i = 0; i < _assets.length; i++) {
+      assets[_assets[i][0]] = _assets[i][1];
+    }
+  })
+  .then(function() {return assets;});
+};
+
+FileSerializer.prototype.write = function(assetOps) {
+  var cacheAssetDirPath = this.path;
+  return mkdirp(cacheAssetDirPath)
+  .then(function() {
+    return assetOps.map(function(asset) {
+      var assetPath = join(cacheAssetDirPath, asset.key);
+      return fsWriteFile(assetPath, asset.value);
+    });
+  })
+  .then(Promise.all);
+};
