+++ lib/hard-source-json-serializer-plugin.js
@@ -1,6 +1,6 @@
 var join = require('path').join;
 
-var JsonSerializer = require('./cache-serializers').JsonSerializer;
+var JsonSerializer = require('./hard-source-json-serializer');
 
 module.exports = HardSourceJsonSerializerPlugin;
 
@@ -10,11 +10,15 @@ HardSourceJsonSerializerPlugin.prototype.apply = function(compiler) {
   compiler.plugin('hard-source-cache-factory', function(factory) {
     return function(info) {
       if (info.type === 'data') {
-        return new JsonSerializer({
-          cacheDirPath: join(info.cacheDirPath, info.name)
-        });
+        return HardSourceJsonSerializerPlugin.createSerializer(info);
       }
       return factory(info);
     };
+  });
+};
+
+HardSourceJsonSerializerPlugin.createSerializer = function(info) {
+  return new JsonSerializer({
+    cacheDirPath: join(info.cacheDirPath, info.name),
   });
 };
